Trait budlang::Environment

pub trait Environment: 'static {
    type String: DynamicValue + for<'a> From<&'a str> + From<String>;
    type Map: DynamicValue + for<'a> TryFrom<PoppedValues<'a>, Error = FaultKind>;
    type List: DynamicValue + FromIterator<Value>;

    fn step(&mut self) -> ExecutionBehavior;

    fn convert(&self, value: &Value, kind: &Symbol) -> Result<Value, FaultKind> { ... }
}

Customizes the behavior of a virtual machine instance.

Required Associated Types

type String: DynamicValue + for<'a> From<&'a str> + From<String>

The string type for this environment.

type Map: DynamicValue + for<'a> TryFrom<PoppedValues<'a>, Error = FaultKind>

The map type for this environment.

type List: DynamicValue + FromIterator<Value>

The list (array) type for this environment.

Required Methods

fn step(&mut self) -> ExecutionBehavior

Called once before each instruction is executed.

If [ExecutionBehavior::Continue] is returned, the next instruction will be exected.

If [ExecutionBehavior::Pause] is returned, the virtual machine is paused and a FaultOrPause::Pause is raised. If the execution is resumed, the first function call will be before executing the same instruction as the one when [ExecutionBehavior::Pause] was called.

Provided Methods

fn convert(&self, value: &Value, kind: &Symbol) -> Result<Value, FaultKind>

Converts value to a custom type supported by the runtime.