Module budlang::vm

The Bud virtual machine.

Structs

Bud

A Bud virtual machine instance.

Budgeted

An Environment that allows executing an amount of instructions before pausing the virtual machine.

Dynamic

A Rust value that has been wrapped for use in the virtual machine.

DynamicFault

A Fault that arose from a Dynamic value.

Fault

An unexpected event occurred while executing the virtual machine.

FaultStackFrame

A stack frame entry of a Fault.

Function

A virtual machine function.

PausedExecution

A paused code execution.

PoppedValues

An iterator over a sequence of values being removed from the top of a Stack.

Stack

A stack of Values.

Enums

CompareAction

An action to take during an Instruction::Compare.

Comparison

A method for comparing Values.

Destination

A destination for a value.

ExecutionBehavior

The virtual machine behavior returned from Environment::step().

FaultKind

An unexpected event within the virtual machine.

FaultOrPause

A reason for a virtual machine Fault.

Instruction

A virtual machine instruction.

Value

A virtual machine value.

ValueKind

All primitive Value kinds.

ValueOrSource

A value or a location of a value

ValueSource

The source of a value.

Traits

DynamicValue

A type that can be used in the virtual machine using Value::dynamic.

Environment

Customizes the behavior of a virtual machine instance.

FromStack

A type that can be constructed from popping from the virtual machine stack.