Trait budlang::vm::Environment

pub trait Environment: 'static {
    fn step(&mut self) -> ExecutionBehavior;
}

Customizes the behavior of a virtual machine instance.

Required Methods

fn step(&mut self) -> ExecutionBehavior

Called once before each instruction is executed.

If ExecutionBehavior::Continue is returned, the next instruction will be exected.

If ExecutionBehavior::Pause is returned, the virtual machine is paused and a FaultKind::Paused is raised. If the execution is resumed, the first function call will be before executing the same instruction as the one when ExecutionBehavior::Pause was called.