Struct budlang::Bud

pub struct Bud<Env>(_)
where
    Env: Environment;

A Bud virtual machine instance.

Each instance of this type has its own sandboxed environment. Its stack space, function declarations, and Environment are unique from all other instances of Bud with the exception that [Symbol]s are tracked globally.

Implementations

impl Bud<()>

pub fn empty() -> Self

Returns a default instance of Bud with no custom Environment

impl<Env> Bud<Env>where
Env: Environment,

pub const fn new(vm: VirtualMachine<BudEnvironment<Env>>) -> Self

Returns a new instance with the provided virtual machine.

pub fn default_for(environment: Env) -> Self

Returns a new instance with the provided environment.

pub fn with_function(self, function: Function<Intrinsic>) -> Self

Registers a function with the provided name and returns self. This is a builder-style function.

pub fn with_native_function(
self,
name: impl Into<Symbol>,
function: impl NativeFunction + 'static
) -> Self

Registers a function with the provided name and returns self. This is a builder-style function.

pub fn evaluate<'a, ReturnType: FromStack>(
&'a mut self,
source: &str
) -> Result<ReturnType, Error<'a, Env, ReturnType>>

Evaluates source interactively and returns the provided result.

Bud is a compiled language. When compiling a chunk of source code, it is organized into a series of declarations. If any non-declaration statements are encountered, they are gathered into an initialization function.

The difference between this function and Bud::run_source() is that the initialization function will be compiled with existing knowledge of any local variables defined in previous code evaluated on this instance. Bud::run_source() always executes the initialization code in its own environment, preventing persisting variables across invoations.

pub fn run_source<Output: FromStack>(
&mut self,
source: &str
) -> Result<Output, Error<'_, Env, Output>>

Compiles source and executes it in this context. Any declarations will persist in the virtual machine, but all local variables will be removed from the stack upon completion.

To enable persisting of local variables, use Bud::evaluate().