Trait budlang::vm::DynamicValue

pub trait DynamicValue: Clone + Debug + Display + 'static {
    fn is_truthy(&self) -> bool;
    fn kind(&self) -> &'static str;

    fn partial_eq(&self, other: &Value) -> Option<bool> { ... }
    fn partial_cmp(&self, other: &Value) -> Option<Ordering> { ... }
    fn call(
        &mut self,
        name: &Symbol,
        args: Drain<'_, Value>
    ) -> Result<Value, FaultKind> { ... }
}

A type that can be used in the virtual machine using Value::dynamic.

Required Methods

fn is_truthy(&self) -> bool

Returns true if the value contained is truthy. See Value::is_truthy() for examples of what this means for primitive types.

fn kind(&self) -> &'static str

Returns a unique string identifying this type. This returned string will be wrapped in ValueKind::Dynamic when Value::kind() is invoked on a dynamic value.

This value does not influence the virtual machine’s behavior. The virtual machine uses this string only when creating error messages.

Provided Methods

fn partial_eq(&self, other: &Value) -> Option<bool>

Returns true if self and other are considered equal. Returns false if self and other are able to be compared but are not equal. Returns None if the values are unable to be compared.

When evaluating left = right with dynamic values, if left.partial_eq(right) returns None, right.partial_eq(left) will be attempted before returning false.

fn partial_cmp(&self, other: &Value) -> Option<Ordering>

Returns the relative ordering of self and other, if a comparison is able to be made. If the types cannot be compared, this function should return None.

When evaluating a comparison such as left < right with dynamic values, if left.partial_cmp(right) returns None, right.partial_cmp(left).map(Ordering::reverse) will be attempted before returning false.

fn call(
&mut self,
name: &Symbol,
args: Drain<'_, Value>
) -> Result<Value, FaultKind>

Calls a function by name with args.